Aggregate function sumMap: use FieldVisitorSum instead of += operator.
